+++ HISTORY.md
@@ -2,6 +2,18 @@ Release History
 ===============
 
 
+Next version
+------------------
+
+**Features and Enhancements**
+
+- Implemented method to check for overlap with genomic region for SVs
+
+**Bugfixes**
+
+- None
+
+
 0.2.0 (2015-03-24)
 ------------------